board/requirements.txt
argcomplete==0.8.1
argh==0.26.1
blessings==1.6
bpython==0.15
chardet==2.3.0
cryptography==1.2.3
curtsies==0.2.6
enum34==1.1.2
greenlet==0.4.9
idna==2.0
ipaddress==1.0.16
kafka-python==1.3.5
ndg-httpsclient==0.4.0
pathtools==0.1.2
pyasn1==0.1.9
Pygments==2.1
pyOpenSSL==0.15.1
PyYAML==3.11
requests==2.9.1
six==1.10.0
urllib3==1.13.1
urwid==1.3.1
watchdog==0.8.3
wcwidth==0.1.6
